[SL-UP] Bugfix: Hardware version persistence. (#223)
