+++ src/platform/silabs/MigrationManager.h
@@ -31,6 +31,7 @@ public:
      */
     static MigrationManager & GetMigrationInstance();
     static void applyMigrations();
+    static void MigrateUint16(uint32_t old_key, uint32_t new_key);
     static void MigrateUint32(uint32_t old_key, uint32_t new_key);
 
 private:
@@ -45,6 +46,7 @@ private:
 void MigrateKvsMap(void);
 void MigrateDacProvider(void);
 void MigrateCounterConfigs(void);
+void MigrateHardwareVersion(void);
 
 } // namespace Silabs
 } // namespace DeviceLayer
